Fixed dependencies
