+++ my-app/frontend/package.json
@@ -1,25 +1,28 @@
 {
-  "name": "my-app",
+  "name": "my-app2",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@chakra-ui/react": "^2.4.3",
-    "@emotion/react": "^11.10.5",
-    "@emotion/styled": "^11.10.5",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "@types/node": "^18.11.15",
-    "express": "^4.18.2",
-    "framer-motion": "^7.8.0",
-    "mongoose": "^6.8.0",
-    "nodemon": "^2.0.20",
-    "passport": "^0.6.0",
+    "@types/jest": "^27.5.2",
+    "@types/node": "^16.18.9",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.9",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
     "typescript": "^4.9.4",
-    "web-vitals": "^2.1.4"
+    "web-vitals": "^2.1.4",
+    "@chakra-ui/react": "^2.4.3",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "express": "^4.18.2",
+    "framer-motion": "^7.8.0",
+    "mongoose": "^6.8.0",
+    "nodemon": "^2.0.20",
+    "passport": "^0.6.0"
   },
   "scripts": {
     "start": "react-scripts start",
