+++ my-app/frontend/src/App.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import "./App.css";
+
+function App() {
+  return <div className="App">Hello World</div>;
+}
+
+export default App;
